src/04_anomaly_detection/ddos/requirements.txt
google-cloud-storage==2.8.0
google-cloud-aiplatform==1.25.0
kfp<2.0
google-cloud-pipeline-components==1.0.42
protobuf==3.20.3
